backend/requirements.txt
Flask
yt-dlp
openai-whisper
Flask-Cors